requirements.txt
Django==5.1.5
Flask==3.1.1
flask_sqlalchemy==3.1.1
Jinja2==3.1.6
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
openpyxl==3.1.5
pandas==2.3.0
pytest==8.3.5
qrcode==8.2
SQLAlchemy==2.0.41
waitress==3.0.2
pillow==11.2.1